Move various RA-specific traits to the new folder/namespace structure
