+++ OpenRA.Mods.RA/Traits/PortableChrono.cs
@@ -10,13 +10,13 @@
 
 using System.Drawing;
 using System.Collections.Generic;
-using OpenRA.Traits;
 using OpenRA.Graphics;
-using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.Common.Orders;
 using OpenRA.Mods.Common.Graphics;
+using OpenRA.Mods.Common.Orders;
+using OpenRA.Mods.RA.Activities;
+using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	class PortableChronoInfo : ITraitInfo
 	{
