+++ OpenRA.Mods.RA/Traits/SupportPowers/GpsPower.cs
@@ -14,7 +14,7 @@ using OpenRA.Effects;
 using OpenRA.Mods.RA.Effects;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	[Desc("Required for GpsPower. Attach this to the player actor.")]
 	class GpsWatcherInfo : ITraitInfo
